change v2

+++ src/main/java/com/pgrg/springbatch/entity/ODSTransactionMessageForChoice.java
@@ -3,6 +3,8 @@ package com.pgrg.springbatch.entity;
 import lombok.*;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
@@ -17,7 +19,7 @@ public class ODSTransactionMessageForChoice {
 
     private String crn;
 
-    private Double totalPointsEarned;
+    private BigDecimal totalPointsEarned;
 
     private String bonusCode;
